requirements.txt
bleak
Pillow
PyYAML
websockets

